SQL_QUERY_GENERATOR/requirements.txt
fastapi
uvicorn
streamlit
pandas
sqlalchemy
pymysql
langchain
langchain-community
langchain-experimental
langchain-groq
langchain-openai
langchain-google-genai
tabulate
requests
transformers
pydantic
openai
sentence-transformers
mysql-connector-python
psycopg2-binary
snowflake-connector-python
tqdm
python-dotenv
loguru
pytest
chromadb